styles and ui fix
test fix

fix import
