+++ client/src/CalendarAddEvent.js
@@ -5,15 +5,18 @@ import {
   TextInput,
   Pressable,
   ScrollView,
+  KeyboardAvoidingView,
   Alert,
-  StyleSheet,
-  Platform
+  Platform,
 } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 import * as Calendar from "expo-calendar";
 import PropTypes from "prop-types";
 import AutocompleteDropdown from "./AutocompleteDropdown"; 
 import { filterLocations, getBuildingDisplayName } from "./utils/locationSearch";
+import styles from "./styles/CalendarAddEvent_styles";
+
+const closeDelay = 150;
 
 function addLead0(n) {
   return String(n).padStart(2, "0");
@@ -43,6 +46,16 @@ function parseDateTime(dateStr, timeStr) {
   return Number.isNaN(d.getTime()) ? null : d;
 }
 
+function formatTimeInput(raw) {
+  const digits = raw.replace(/\D/g, "").slice(0, 4);
+  if (digits.length < 4) return digits;
+  const firstTwo = Number.parseInt(digits.slice(0, 2), 10);
+  if (firstTwo > 23) {
+    return `0${digits[0]}:${digits.slice(1, 3)}`;
+  }
+  return `${digits.slice(0, 2)}:${digits.slice(2)}`;
+}
+
 export default function CalendarAddEvent({
   selectedCalendarIds,
   onEventAdded,
@@ -67,7 +80,7 @@ export default function CalendarAddEvent({
   const locationResults = activeField === "location" ? filterLocations(locationInput, []) : [];
 
   const scheduleClose = () => {
-    setTimeout(() => setActiveField((prev) => (prev === "location" ? null : prev)), 150);
+    setTimeout(() => setActiveField((prev) => (prev === "location" ? null : prev)), closeDelay);
   };
 
   const pickLocation = (loc) => {
@@ -87,8 +100,8 @@ export default function CalendarAddEvent({
       Alert.alert("Missing title", "Please enter an event title.");
       return false;
     }
-    const start = parseDateTime(dateStr, startTimeStr);
-    const end = parseDateTime(dateStr, endTimeStr);
+    const start = parseDateTime(dateStr, formatTimeInput(startTimeStr));
+    const end = parseDateTime(dateStr, formatTimeInput(endTimeStr));
     if (!start) {
       Alert.alert("Invalid date/time", "Please check the date and start time.");
       return false;
@@ -138,6 +151,10 @@ export default function CalendarAddEvent({
   };
 
   return (
+    <KeyboardAvoidingView
+      style={styles.container}
+      behavior={Platform.OS === "ios" ? "padding" : "height"}
+    >
     <ScrollView
       style={styles.container}
       contentContainerStyle={styles.content}
@@ -180,10 +197,10 @@ export default function CalendarAddEvent({
             <Text style={styles.timeLabel}>Start</Text>
             <TextInput
               style={[styles.input, styles.timeInput]}
-              placeholder="HH:MM"
+              placeholder="HH:MM, 24h"
               placeholderTextColor="#aaa"
               value={startTimeStr}
-              onChangeText={setStartTimeStr}
+              onChangeText={(v) => setStartTimeStr(formatTimeInput(v))}
               keyboardType="numeric"
               maxLength={5}
               returnKeyType="done"
@@ -196,10 +213,10 @@ export default function CalendarAddEvent({
             <Text style={styles.timeLabel}>End</Text>
             <TextInput
               style={[styles.input, styles.timeInput]}
-              placeholder="HH:MM"
+              placeholder="HH:MM, 24h"
               placeholderTextColor="#aaa"
               value={endTimeStr}
-              onChangeText={setEndTimeStr}
+              onChangeText={(v) => setEndTimeStr(formatTimeInput(v))}
               keyboardType="numeric"
               maxLength={5}
               returnKeyType="done"
@@ -265,6 +282,7 @@ export default function CalendarAddEvent({
         </View>
       </View>
     </ScrollView>
+    </KeyboardAvoidingView>
   );
 }
 
@@ -273,114 +291,3 @@ CalendarAddEvent.propTypes = {
   onEventAdded: PropTypes.func.isRequired,
   onCancel: PropTypes.func.isRequired,
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#f5f5f5",
-  },
-  content: {
-    paddingBottom: 40,
-  },
-  header: {
-    backgroundColor: "#912338",
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    paddingTop: Platform.OS === "ios" ? 56 : 40,
-    paddingBottom: 16,
-    paddingHorizontal: 16,
-  },
-  headerBtn: {
-    width: 36,
-  },
-  headerTitle: {
-    color: "white",
-    fontSize: 18,
-    fontWeight: "600",
-  },
-  form: {
-    padding: 20,
-    gap: 4,
-  },
-  label: {
-    fontSize: 13,
-    fontWeight: "600",
-    color: "#555",
-    marginTop: 16,
-    marginBottom: 4,
-    textTransform: "uppercase",
-    letterSpacing: 0.5,
-  },
-  input: {
-    backgroundColor: "white",
-    borderRadius: 10,
-    paddingHorizontal: 14,
-    paddingVertical: 12,
-    fontSize: 15,
-    color: "#222",
-    borderWidth: 1,
-    borderColor: "#e0e0e0",
-  },
-  timeRow: {
-    flexDirection: "row",
-    alignItems: "flex-end",
-    gap: 8,
-  },
-  timeFieldWrap: {
-    flex: 1,
-  },
-  timeLabel: {
-    fontSize: 12,
-    color: "#888",
-    marginBottom: 4,
-  },
-  timeInput: {
-    textAlign: "center",
-  },
-  timeArrow: {
-    marginBottom: 14,
-  },
-  locationWrap: {
-    position: "relative",
-  },
-  roomInput: {
-    marginTop: 8,
-  },
-  clearBtn: {
-    marginLeft: 8,
-  },
-  actions: {
-    flexDirection: "row",
-    gap: 12,
-    marginTop: 32,
-  },
-  cancelBtn: {
-    flex: 1,
-    paddingVertical: 14,
-    borderRadius: 10,
-    borderWidth: 1,
-    borderColor: "#912338",
-    alignItems: "center",
-  },
-  cancelTxt: {
-    color: "#912338",
-    fontSize: 15,
-    fontWeight: "600",
-  },
-  saveBtn: {
-    flex: 2,
-    paddingVertical: 14,
-    borderRadius: 10,
-    backgroundColor: "#912338",
-    alignItems: "center",
-  },
-  saveBtnDisabled: {
-    opacity: 0.6,
-  },
-  saveTxt: {
-    color: "white",
-    fontSize: 15,
-    fontWeight: "600",
-  },
-});
